requirements.txt
asgiref==3.7.2
Django==4.2.9
sqlparse==0.4.4
psycopg2==2.9.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
